Revert "Refactor: Rename bundleidentifier to needforheat"
This reverts commit a59277117c62589a80861f4676322d1b68cb71e2.
