+++ apps/expo/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "nl.windesheim.energietransitie.needforheat",
+    package = "nl.windesheim.energietransitie.warmtewachter",
 )
 
 android_resource(
     name = "res",
-    package = "nl.windesheim.energietransitie.needforheat",
+    package = "nl.windesheim.energietransitie.warmtewachter",
     res = "src/main/res",
 )
