+++ apps/expo/android/app/build.gradle
@@ -143,7 +143,7 @@ android {
     compileSdkVersion rootProject.ext.compileSdkVersion
 
     defaultConfig {
-        applicationId 'nl.windesheim.energietransitie.needforheat'
+        applicationId 'nl.windesheim.energietransitie.warmtewachter'
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 20026
