+++ apps/expo/android/app/src/debug/java/nl/windesheim/needforheat/ReactNativeFlipper.java
@@ -4,7 +4,7 @@
  * <p>This source code is licensed under the MIT license found in the LICENSE file in the root
  * directory of this source tree.
  */
-package nl.windesheim.energietransitie.needforheat;
+package nl.windesheim.energietransitie.warmtewachter;
 
 import android.content.Context;
 import com.facebook.flipper.android.AndroidFlipperClient;
