+++ apps/expo/android/app/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="nl.windesheim.energietransitie.needforheat">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="nl.windesheim.energietransitie.warmtewachter">
   <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
   <uses-permission android:name="android.permission.BLUETOOTH"/>
   <uses-permission android:name="android.permission.BLUETOOTH_ADMIN"/>
