+++ apps/expo/android/app/src/main/java/nl/windesheim/energietransitie/warmtewachter/MainActivity.java
@@ -1,4 +1,4 @@
-package nl.windesheim.energietransitie.needforheat;
+package nl.windesheim.energietransitie.warmtewachter;
 
 import android.os.Build;
 import android.os.Bundle;
