+++ apps/expo/android/app/src/main/java/nl/windesheim/energietransitie/warmtewachter/MainApplication.java
@@ -1,4 +1,4 @@
-package nl.windesheim.energietransitie.needforheat;
+package nl.windesheim.energietransitie.warmtewachter;
 
 import android.app.Application;
 import android.content.Context;
@@ -12,7 +12,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.config.ReactFeatureFlags;
 import com.facebook.soloader.SoLoader;
-import nl.windesheim.energietransitie.needforheat.newarchitecture.MainApplicationReactNativeHost;
+import nl.windesheim.energietransitie.warmtewachter.newarchitecture.MainApplicationReactNativeHost;
 
 import expo.modules.ApplicationLifecycleDispatcher;
 import expo.modules.ReactNativeHostWrapper;
@@ -92,7 +92,7 @@ public class MainApplication extends Application implements ReactApplication {
          * We use reflection here to pick up the class that initializes Flipper,
          * since Flipper library is not available in release mode
          */
-        Class<?> aClass = Class.forName("nl.windesheim.energietransitie.needforheat.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("nl.windesheim.energietransitie.warmtewachter.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
