+++ apps/expo/android/app/src/main/java/nl/windesheim/energietransitie/warmtewachter/newarchitecture/components/MainComponentsRegistry.java
@@ -1,4 +1,4 @@
-package nl.windesheim.energietransitie.needforheat.newarchitecture.components;
+package nl.windesheim.energietransitie.warmtewachter.newarchitecture.components;
 
 import com.facebook.jni.HybridData;
 import com.facebook.proguard.annotations.DoNotStrip;
