+++ apps/expo/android/app/src/main/java/nl/windesheim/energietransitie/warmtewachter/newarchitecture/modules/MainApplicationTurboModuleManagerDelegate.java
@@ -1,4 +1,4 @@
-package nl.windesheim.energietransitie.needforheat.newarchitecture.modules;
+package nl.windesheim.energietransitie.warmtewachter.newarchitecture.modules;
 
 import com.facebook.jni.HybridData;
 import com.facebook.react.ReactPackage;
